requirements.txt
pycryptodome
wxpython
pillow
